# docs/swagger.yaml
definitions:
  dto.AdminLoginInput:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 账号
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
    required:
    - password
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: 管理员信息
      operationId: /admin/admin_info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 管理员信息
      tags:
      - 管理员接口
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 修改密码
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 修改密码
      tags:
      - 管理员接口
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 管理员登陆
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 管理员登陆
      tags:
      - 管理员接口
  /admin_login/logout:
    get:
      consumes:
      - application/json
      description: 管理员退出
      operationId: /admin_login/logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 管理员退出
      tags:
      - 管理员接口
  /service/service_list:
    get:
      consumes:
      - application/json
      description: 服务列表
      operationId: /service/service_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 服务列表
      tags:
      - 服务管理
swagger: "2.0"